requirements.txt
aiogram==3.15.0
yt-dlp>=2024.12.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
mutagen==1.47.0
aiofiles==24.1.0
